Update to Sept 28th upstream commit.
Attempt to include the new tracy profiler support, something is wrong at the cmake configure step so that it fails right after tracy is successfully built and installed.  This may be the way tracy is called for the build of minetest, not sure how to fix but it may be a cmake variable, or similar.